mvp/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
PyYAML==6.0.2
python-multipart==0.0.12
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
requests==2.32.3
